# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Boilerplate",
    "summary": "A modular FastAPI starter with a plugin system",
    "description": "\n    # FastAPI Boilerplate\n\n    A modern FastAPI starter with:\n\n    * Vertical-slice modules and a clean infrastructure layer\n    * Session-based auth with OAuth providers\n    * Swappable cache, queue, and rate-limit backends\n    * SQLAdmin admin UI\n    ",
    "contact": {
      "name": "Benav Labs",
      "url": "https://github.com/benavlabs/FastAPI-boilerplate",
      "email": "contact@benav.io"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.18.0"
  },
  "paths": {
    "/api/v1/users/": {
      "post": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "Create New User Account",
        "description": "Creates a new user account in the system.\n\n           This endpoint allows registration of new users with their basic information:\n           - Full name\n           - Username (must be lowercase alphanumeric)\n           - Email address\n           - Password (with security requirements)\n\n           New accounts are automatically assigned to the default tier.",
        "operationId": "create_user_api_v1_users__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user data"
          },
          "409": {
            "description": "Username or email already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "Get Users",
        "description": "Retrieves a paginated list of all users in the system.\n\n           This admin-only endpoint provides access to all user accounts and supports\n           pagination to handle large numbers of users efficiently. The results include\n           basic profile information for each user.\n\n           For security reasons, sensitive information like passwords is never included\n           in the response.",
        "operationId": "get_users_api_v1_users__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Items Per Page"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of users with total count and pagination metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedData_UserRead_"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized - requires admin privileges"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "Get Current User Profile",
        "description": "Retrieves the profile information of the currently authenticated user.\n\n            This endpoint provides users with their own profile data including:\n            - Basic profile information (name, username, email)\n            - Profile image URL\n            - Subscription tier information\n            - Authentication details (superuser status, email verification)\n\n            This is a convenient way for frontend applications to get the current\n            user's information for display or personalization purposes.",
        "operationId": "get_current_user_profile_api_v1_users_me_get",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The current user's profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/active-and-inactive/{username}": {
      "get": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "Get Active and Inactive User Profile by Username(Admin)",
        "description": "Retrieve a user's profile information by their unique username.\n\n            This endpoint can be used to look up any user in the system by their\n            username. It returns the same profile data structure as other user\n            endpoints but does not include sensitive information.\n\n            Note that usernames are case-sensitive in lookup operations.",
        "operationId": "get_active_and_inactive_user_by_username_api_v1_users_active_and_inactive__username__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the requested user's profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized - requires admin privileges"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/db/{username}": {
      "delete": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "GDPR Delete User (Admin)",
        "description": "GDPR/LGPD compliant user deletion with data anonymization.\n\n            This admin-only endpoint anonymizes user PII while preserving business data\n            integrity and maintaining referential relationships for conversations and\n            analytics data.\n\n            This operation:\n            - Removes personally identifiable information (PII)\n            - Retains email for legal compliance purposes\n            - Prevents future login by clearing credentials\n            - Maintains foreign key relationships (conversations, logs)\n            - Logs the deletion event with legal basis for audit compliance\n\n            Unlike hard deletion, this approach:\n            - Complies with GDPR Article 17 (Right to Erasure)\n            - Preserves business analytics data\n            - Eliminates foreign key constraint violations\n            - Maintains audit trails for legal requirements\n\n            This operation is needed for:\n            - GDPR/LGPD data deletion requests\n            - Legal compliance while preserving business data\n            - Safe user removal without breaking referential integrity",
        "operationId": "gdpr_delete_user_api_v1_users_db__username__delete",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User anonymized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Gdpr Delete User Api V1 Users Db  Username  Delete"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized - requires admin privileges"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 30,
            "minLength": 2,
            "title": "Name",
            "examples": [
              "User Userson"
            ]
          },
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 2,
            "pattern": "^[a-z0-9]+$",
            "title": "Username",
            "examples": [
              "userson"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user.userson@example.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "pattern": "^.{8,}|[0-9]+|[A-Z]+|[a-z]+|[^a-zA-Z0-9]+$",
            "title": "Password",
            "description": "Password must be at least 8 characters long and include a number,uppercase letter, lowercase letter, and special character",
            "examples": [
              "Str1ngst!"
            ]
          },
          "google_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Id"
          },
          "github_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Id"
          },
          "oauth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Provider"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "default": false
          },
          "oauth_created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Created At"
          },
          "oauth_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "Schema for creating a new user."
      },
...
    }
  }
}